+++ main.go
@@ -2,19 +2,29 @@ package main
 
 import (
 	"fmt"
-	"github.com/jrswab/go-status/blocks" // if pulled from github change to github
-	"github.com/jrswab/go-status/ui"     // if pulled from github change to github
+	"github.com/jrswab/go-status/blocks"
+	"github.com/jrswab/go-status/ui"
 	"log"
-	"math"
 	"time"
 )
 
+// To Add Battery Status:
+// Uncomment all lines that follow the message:
+// "Uncomment for battery status"
+// Also, add the correct string formatting to
+// the status assignment at the end of the file and add "math" to the imports
+
 func main() {
-	var status, hTime, battery string
-	var weather, ramFree, homeSpace string                        // Go routine blocks
-	var wttrErr, ramErr, homeErr error                            // Go routine errors
-	var rShift, dropbox, volIcon, volText, syncthing, bolt string // Icon blocks
-	var passed, fiveMins float64
+	var weather, ramFree, homeSpace string // Go routine blocks
+	var wttrErr, ramErr, homeErr error     // Go routine errors
+	var status, hTime, rShift, dropbox, volIcon, volText, syncthing string
+
+	// Uncomment for battery status
+	/*
+		var bolt, battery string
+		var passed, fiveMins float64
+		start := time.Now()
+	*/
 
 	// These are static icons and only need defined at the start
 	homeDir := blocks.DirIcon()
@@ -35,12 +45,15 @@ func main() {
 	go blocks.HomeDisk(cHomeDisk, eHomeDisk)
 
 	ticker := time.NewTicker(time.Second)
-	start := time.Now()
 	for range ticker.C {
 		// add year & seconds with "Jan 02, 2006 15:04:05"
 		hTime = time.Now().Format("Jan 02 15:04")
-		passed = time.Since(start).Seconds()
-		fiveMins = math.Floor(math.Remainder(passed, 300))
+
+		// Uncomment for battery status
+		/*
+			passed = time.Since(start).Seconds()
+			fiveMins = math.Floor(math.Remainder(passed, 300))
+		*/
 
 		select { // update the go routine channels as they send data
 		case weather = <-cWttr:
@@ -61,15 +74,19 @@ func main() {
 		volText, _ = blocks.VolumeText()
 		volIcon, _ = blocks.VolumeIcon()
 		syncthing, _ = blocks.SyncthingIcon()
-		bolt = blocks.PowerIcon()
-		if fiveMins == 0 || passed < 10 {
-			battery, _ = blocks.Battery()
-		}
+
+		// Uncomment for battery status
+		/*
+			bolt = blocks.PowerIcon()
+			if fiveMins == 0 || passed < 10 {
+				battery, _ = blocks.Battery()
+			}
+		*/
 
 		// Change by editing variables & `%s`
-		status = fmt.Sprintf(" %s%s %s%s %s%s %s%s %s%s %s %s%s%s",
+		status = fmt.Sprintf(" %s%s %s%s %s%s %s%s %s %s%s%s",
 			tempIco, weather, homeDir, homeSpace, memIco, ramFree, volIcon, volText,
-			bolt, battery, hTime, dropbox, syncthing, rShift)
+			hTime, dropbox, syncthing, rShift)
 		ui.Dwm(status) // change this to the UI of choice
 	}
 }