+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- In code instructions to enable battery status.
+
+### Changed
+- Commented out all battery status code in main.go
 
 ## [0.8.1] - July 3, 2019
 ## Changed