+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- In code instructions to enable battery status.
+- Icons package for clearer code.
+
+### Changed
+- Commented out all battery status code in main.go
+- wttr custom error message to used fmt.Errorf instead for errors.New
+- Icons from calls like `blocks.VolumeIcon()` to calls like `icons.Volume()`.
+
+### Removed
+- The errors package from wttr.go
 
 ## [0.8.1] - July 3, 2019
 ## Changed